Correção de problemas:
1. Defensive Midfielder v1.0
2. Defender v1.1

+++ entity/player/behaviour/basics/behaviour_cover.h
@@ -35,7 +35,7 @@ private:
 
     // Parameters
     int _id;
-    int _a, _b, _c;
+    float _a, _b, _c;
 
 public:
     Behaviour_Cover();
